test: GamerTest 케이스 추가
