+++ src/test/java/blackjack/domain/gamer/DealerTest.java
@@ -4,26 +4,24 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class DealerTest {
 
-    private Dealer dealer = new Dealer();
+    @ParameterizedTest
+    @CsvSource(value = {"14:true", "16:false"}, delimiter = ':')
+    @DisplayName("보유 카드 번호 합이 특정 숫자를 넘었는지 확인")
+    void checkCardsNumberSum(int input, boolean result) {
+        Dealer dealer = new Dealer();
+        Card card1 = Card.getInstance(CardShape.SPADE, CardNumber.TEN);
+        Card card2 = Card.getInstance(CardShape.SPADE, CardNumber.FIVE);
 
-    @Test
-    @DisplayName("입력 받은 값 보다 크면 true를 반환한다.")
-    void isOverThanTrue() {
-        dealer.addCard(Card.getInstance(CardShape.CLOVER, CardNumber.ACE));
-        assertThat(dealer.isOverThan(10)).isTrue();
-    }
+        dealer.addCard(card1);
+        dealer.addCard(card2);
 
-    @Test
-    @DisplayName("입력 받은 값 보다 작으면 false를 반환한다.")
-    void isOverThanFalse() {
-        dealer.addCard(Card.getInstance(CardShape.CLOVER, CardNumber.ACE));
-        assertThat(dealer.isOverThan(11)).isFalse();
+        assertThat(dealer.isOverThan(input)).isEqualTo(result);
     }
 }
