+++ src/test/java/blackjack/domain/gamer/GamerTest.java
@@ -1,7 +1,6 @@
 package blackjack.domain.gamer;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.CardFactory;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
 import org.junit.jupiter.api.DisplayName;
@@ -28,20 +27,6 @@ class GamerTest {
         assertThat(cards.size()).isEqualTo(1);
     }
 
-    @ParameterizedTest
-    @CsvSource(value = {"14:true", "16:false"}, delimiter = ':')
-    @DisplayName("보유 카드 번호 합이 특정 숫자를 넘었는지 확인")
-    void checkCardsNumberSum(int input, boolean result) {
-        Dealer dealer = new Dealer();
-        Card card1 = Card.getInstance(CardShape.SPADE, CardNumber.TEN);
-        Card card2 = Card.getInstance(CardShape.SPADE, CardNumber.FIVE);
-
-        dealer.addCard(card1);
-        dealer.addCard(card2);
-
-        assertThat(dealer.isOverThan(input)).isEqualTo(result);
-    }
-
     @Test
     @DisplayName("보유 카드 번호 합 반환")
     void calculateCardsNumberSum() {
@@ -57,9 +42,8 @@ class GamerTest {
     }
 
     @Test
-    @DisplayName("Ace 포함된 보유 카드 번호 합 반환")
-    void calculateCardsNumberSumWithAce() {
-        Gamer gamer = new Gamer("name");
+    @DisplayName("Ace가 1로 계산된 보유 카드 번호 합 반환")
+    void calculateCardsNumberSumWithAceOne() {
         Card card1 = Card.getInstance(CardShape.SPADE, CardNumber.NINE);
         Card card2 = Card.getInstance(CardShape.SPADE, CardNumber.QUEEN);
         Card card3 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
@@ -70,5 +54,35 @@ class GamerTest {
 
         int sum = gamer.getCardsNumberSum();
         assertThat(sum).isEqualTo(20);
+    }
+
+    @Test
+    @DisplayName("Ace가 11로 계산된 보유 카드 번호 합 반환")
+    void calculateCardsNumberSumWithAceEleven() {
+        Card card1 = Card.getInstance(CardShape.SPADE, CardNumber.JACK);
+        Card card2 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
+
+        gamer.addCard(card1);
+        gamer.addCard(card2);
+
+        int sum = gamer.getCardsNumberSum();
+        assertThat(sum).isEqualTo(21);
+    }
+
+    @Test
+    @DisplayName("Ace가 4장일 때 보유 카드 번호 합 반환")
+    void calculateCardsNumberSumWithFourAce() {
+        Card card1 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
+        Card card2 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
+        Card card3 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
+        Card card4 = Card.getInstance(CardShape.SPADE, CardNumber.ACE);
+
+        gamer.addCard(card1);
+        gamer.addCard(card2);
+        gamer.addCard(card3);
+        gamer.addCard(card4);
+
+        int sum = gamer.getCardsNumberSum();
+        assertThat(sum).isEqualTo(14);
     }
 }
